+++ core/src/main/scala/ch/epfl/scala/debugadapter/internal/scalasig/Decompiler.scala
@@ -1,4 +1,6 @@
-package ch.epfl.scala.debugadapter.internal
+package ch.epfl.scala.debugadapter.internal.scalasig
+
+import ch.epfl.scala.debugadapter.internal.ClassFile
 
 import java.lang.StringBuilder
 import java.nio.charset.StandardCharsets.UTF_8
@@ -9,13 +11,11 @@ import org.objectweb.asm.ClassVisitor
 import org.objectweb.asm.AnnotationVisitor
 import org.objectweb.asm.Opcodes
 
-import ch.epfl.scala.debugadapter.internal.scalasig._
-
 /**
  * Originally copied from https://github.com/JetBrains/intellij-scala
  * https://github.com/JetBrains/intellij-scala/blob/074e8f98d9789b3e7def3ade8d39e7ae770beccf/scala/decompiler/src/org/jetbrains/plugins/scala/decompiler/scalasig/Decompiler.scala
  */
-object Decompiler {
+private[internal] object Decompiler {
   private val BYTES_VALUE = "bytes"
   private val SCALA_SIG_ANNOTATION: String = "Lscala/reflect/ScalaSignature;"
   private val SCALA_LONG_SIG_ANNOTATION: String =
